backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
dashscope>=1.14.0
requests>=2.31.0
websockets>=12.0
aiofiles>=23.2.1
pypinyin>=0.50.0
